Fixes issues raised from feedback
Changes <button> elements to be created with the type attribute set to
'button'.

Adds the option to choose between background image and 'inline' image for
the optional personality image on the result screen.

Adds an option to skip the start screen jumping straight to the first
question of the quiz.

Fixes a bug where the quiz would crash if one or more but not all answers
had an image associated with them.

Changes the way responsiveness is managed. Removes the css media query and
adds a threshold for columns. If the threshold is crossed the column
layout is recalculated.

Prefixes all css class names to avoid collisions with the surrounding
environment.

Adds documentation to all functions.

Adds and fixes description for several fields in semantics.json.