Fixes a couple of styling issues and changes the embedType to iframe
Changes embedType to iframe because of unfortunate interactions between
the quiz styling and the parent page.

Changes the image answer buttons because of an issue where the buttons
would not fill the parent element correctly.

Simplifies the inline result image calculation hopefully removing an issue
where the image height would be set to 0px.